+++ src/inspect_ai/scorer/_metrics/__init__.py
@@ -1,4 +1,5 @@
 from .accuracy import accuracy
+from .aggregate import aggregate
 from .grouped import grouped
 from .mean import mean
 from .perplexity import perplexity_per_seq, perplexity_per_token
@@ -6,6 +7,7 @@ from .std import bootstrap_stderr, std, stderr, var
 
 __all__ = [
     "accuracy",
+    "aggregate",
     "mean",
     "grouped",
     "perplexity_per_token",